fix compilation bugs in CMSSW_6_2_4

+++ TagAndProbeTreeWriter/src/IsoTreeWriter.cc
@@ -262,9 +262,9 @@ double IsoTreeWriter<T>::calcIsoMinPt(const pat::Electron& lepton)
 {
 	double value = -1.0;
 	if (lepton.isEB())
-		value = (lepton.dr03HcalTowerSumEt()+lepton.dr03EcalRecHitSumEt()+lepton.dr03TkSumPt())/std::max(lepton.pt(), 20.0);
+		value = (lepton.dr03HcalTowerSumEt()+lepton.dr03EcalRecHitSumEt()+lepton.dr03TkSumPt())/std::max(lepton.pt(), (float) 20.0);
 	else
-		value = (lepton.dr03HcalTowerSumEt() + std::max(0.0, lepton.dr03EcalRecHitSumEt() - 1.0) + lepton.dr03TkSumPt())/std::max(lepton.pt(), 20.0);
+		value = (lepton.dr03HcalTowerSumEt() + std::max(0.0, lepton.dr03EcalRecHitSumEt() - 1.0) + lepton.dr03TkSumPt())/std::max(lepton.pt(), (float) 20.0);
 
 	return value;
 }
